mini_projects/kiss.daniel/3. házi feladat/pyproject.toml
[tool.poetry]
name = "ai-weather-agent"
version = "0.1.0"
description = "CLI-based AI weather agent using LangGraph and Ollama"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain-core = "^0.3.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
langchain-ollama = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
weather-agent = "src.main:main"
